+++ ui/tests/integration/components/list-test.js
@@ -5,12 +5,13 @@
 
 import { module, test } from 'qunit';
 import { setupRenderingTest } from 'vault/tests/helpers';
-import { render, click } from '@ember/test-helpers';
+import { render, click, find, findAll } from '@ember/test-helpers';
 import { hbs } from 'ember-cli-htmlbars';
 import { v4 as uuidv4 } from 'uuid';
 import sinon from 'sinon';
 import { setupMirage } from 'ember-cli-mirage/test-support';
 import { overrideResponse } from 'vault/tests/helpers/stubs';
+import { clickTrigger } from 'ember-power-select/test-support/helpers';
 
 import { createSecretsEngine } from 'vault/tests/helpers/secret-engine/secret-engine-helpers';
 import { SECRET_ENGINE_SELECTORS as SES } from 'vault/tests/helpers/secret-engine/secret-engine-selectors';
@@ -19,17 +20,6 @@ import { GENERAL } from 'vault/tests/helpers/general-selectors';
 module('Integration | Component | secret-engine/list', function (hooks) {
   setupRenderingTest(hooks);
   setupMirage(hooks);
-
-  // Incoming tests:
-  // allows you to disable an engine
-  // adds disabled css styling to unsupported secret engines
-  // filters by name and engine type
-  // applies overflow styling
-
-  // Add these tests:
-  // shows only displayable engines (whatever that means)
-  // sorts the engines first by supported and then name
-  //
 
   hooks.beforeEach(function () {
     this.server.post('/sys/capabilities-self', () => ({
@@ -48,6 +38,8 @@ module('Integration | Component | secret-engine/list', function (hooks) {
     this.secretEngineModels = [
       createSecretsEngine(this.store, 'cubbyhole', 'cubbyhole-test'),
       createSecretsEngine(this.store, 'kv', 'kv2-test'),
+      createSecretsEngine(this.store, 'aws', 'aws-1'),
+      createSecretsEngine(this.store, 'aws', 'aws-2'),
       createSecretsEngine(this.store, 'nomad', 'nomad-test'),
     ];
   });
@@ -63,12 +55,58 @@ module('Integration | Component | secret-engine/list', function (hooks) {
     assert.dom(SES.secretsBackendLink(enginePath)).exists('shows the link for the kvv2 secrets engine');
     const row = SES.secretsBackendLink(enginePath);
     await click(`${row} ${GENERAL.menuTrigger}`);
-    await click(`${row} ${GENERAL.confirmTrigger}`);
+    await click(GENERAL.menuItem('disable-engine'));
     await click(GENERAL.confirmButton);
 
     assert.true(
       this.flashSuccessSpy.calledWith(`The kv Secrets Engine at ${enginePath}/ has been disabled.`),
       'Flash message shows that engine was disabled.'
     );
+  });
+
+  test('it adds disabled css styling to unsupported secret engines', async function (assert) {
+    await render(hbs`<SecretEngine::List @secretEngineModels={{this.secretEngineModels}} />`);
+    assert
+      .dom(SES.secretsBackendLink('nomad-test'))
+      .doesNotHaveClass(
+        'linked-block',
+        `the linked-block class is not added to the unsupported nomad engine, which effectively disables it.`
+      );
+
+    assert
+      .dom(SES.secretsBackendLink('aws-1'))
+      .hasClass('linked-block', `linked-block class is added to supported aws engines.`);
+  });
+
+  //
+  test('it filters by name and engine type', async function (assert) {
+    await render(hbs`<SecretEngine::List @secretEngineModels={{this.secretEngineModels}} />`);
+    // filter by type
+    await clickTrigger('#filter-by-engine-type');
+    await click(GENERAL.searchSelect.option());
+
+    const rows = findAll(SES.secretsBackendLink());
+    const rowsAws = Array.from(rows).filter((row) => row.innerText.includes('aws'));
+
+    assert.strictEqual(rows.length, rowsAws.length, 'all rows returned are aws');
+    // filter by name
+    await clickTrigger('#filter-by-engine-name');
+    const firstItemToSelect = find(GENERAL.searchSelect.option()).innerText;
+    await click(GENERAL.searchSelect.option());
+    const singleRow = document.querySelectorAll(SES.secretsBackendLink());
+    assert.strictEqual(singleRow.length, 1, 'returns only one row');
+    assert.dom(singleRow[0]).includesText(firstItemToSelect, 'shows the filtered by name engine');
+
+    // clear filter by engine name
+    await click(`#filter-by-engine-name ${GENERAL.searchSelect.removeSelected}`);
+    const rowsAgain = document.querySelectorAll(SES.secretsBackendLink());
+    assert.ok(rowsAgain.length > 1, 'filter has been removed');
+  });
+
+  test('it applies overflow styling', async function (assert) {
+    await render(hbs`<SecretEngine::List @secretEngineModels={{this.secretEngineModels}} />`);
+    // not using the secret-engine-selector "secretPath" because I want to return the first node of a querySelectorAll
+    const firstSecretEngine = document.querySelectorAll('[data-test-secret-path]')[0];
+    assert.dom(firstSecretEngine).hasClass('overflow-wrap', 'secret engine name has overflow class ');
   });
 });
